+++ atat/src/asynch/mod.rs
@@ -1,7 +1,5 @@
-mod buffers;
 mod client;
 mod ingress;
 
-pub use buffers::Buffers;
 pub use client::{AtatClient, Client};
 pub use ingress::{AtatIngress, Ingress};
